Added the first bash integration
Its now possible to source 1 file and everything should just work.
Also improve the environment variable generation. The idea has changed a
few times, but I have ended up with thinking that storing the
environment setup in a json file and generate scripts on the fly is the
best approach. Also adding this information to the buildset files makes
them very build dependent with hard paths etc, hence storing them in a
seperate file in $build_dir/build_shell/environment.json. This is work
in progress, and this commit does not build