Fail when finding a script that should exist fails
ie. all phases should find a script. However, phases like pre and post
build shouldn't fail if we can find a script for them.

Also fix the fact that we where messing around with 1 Process instance
in handle build. This failed when we changed the fallback identifier
when deepcleaning. Added a very crude ProcessBuilder so that each phase
has its own Process instance