Added support for multiple repositories for one project repository
This is handy when building clang where you first clone llvm and then
into a subdirectory of llvm you clone clang and compiler-rt

The problem is what to do on generation. Should I go through all
directories and look for a .git folder, or should I only update the one
which already exist in the buildset file. This is not implemented yet